Drop case normalization from tags and attributes
We trust the user to use proper capitalization for element names and
attributes. Most of the time this doesn't matter, except when htmlbars
or another engine wants to generate DOM using the tokens. The odd
capitalization rules (like the viewBox attribute name, or linearGradient
tag name) must be permitted to pass through.

The alternative approach, requiring that downstream libraries
re-normalize viewbox to viewBox, seems difficult to maintain
practically. Trusting the user to do what is right is simpler.