Fix location information for unquoted attributes.
When an unquoted attribute is parsed we were eagerly consuming the
character.  This meant that when `finishAttributeValue` was called the
`this.column` and `this.line` have been incremented to include the next
character.

This changes unquoted attributes to peek first, then consume when a
given value is categorized. This is not an issue for quoted attributes
because we are always consuming up to and including the final quote
character.