Track attribute and attribute value start information.
This makes it much easier for downstream consumers to deal with
attribute location information.

Currently, in HTMLBars and Glimmer we have to grab the start information
manually during `beginAttribute`, but I would much prefer to simply
track that information when we finish the attribute.

This is roughly inline with what is done for `tagLine` and `tagColumn`.