tokenize() options should be optional
The tokenize() helper function takes a mandatory second argument of options that are passed on to the Tokenizer constructor. In the constructor, however, the options hash is optional, making the helper function more strict than the constructor it delegates to. This commit changes the function signature of tokenize() to make the options hash optional. It also adds type checking to the options argument, which was implicit `any` before.