Introduce public `transitionTo` to change states
This PR cleans up some types and also introduces a public API for transitioning the state
of the EventTokenizer. In the Glimmer VM we are reaching directly mutating the state field
which seems less than ideal. We also calling methods on the delegate, thus it to has been
made public.