Fix regression with unquoted attributes without space before self-closing tag.
The prior change (adding `'/'` support to `afterAttributeUnquoted`)
missed an important step: you have to
`this.delegate.finishAttributeValue()` before transitioning out of
`afterAttributeUnquoted` state (into `selfClosingStartTag` state).

The tests added along with that change passed due to the test harness
(and non-evented tokenizer) being implemented in a way that completely
ignores `finishAttributeValue` hook.

This change fixes the regression and updates the non-evented tokenizer
to more properly match the pattern implemented in glimmer-vm (where
`finishAttributeValue` actually appends the attribute to the current
element node).